+++ src/IntegrationTests/Tests/Commands/Shipments/CreateShipmentComandTests.cs
@@ -39,7 +39,7 @@ public sealed class CreateShipmentComandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
@@ -69,7 +69,7 @@ public sealed class CreateShipmentComandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
@@ -102,7 +102,7 @@ public sealed class CreateShipmentComandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1, 1, 1 },
+            ShipmentParts = new List<double> { 1, 1, 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
