+++ src/IntegrationTests/Tests/Commands/Shipments/SubmitShipmentCommandTests.cs
@@ -39,7 +39,7 @@ public sealed class SubmitShipmentCommandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
@@ -81,7 +81,7 @@ public sealed class SubmitShipmentCommandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
@@ -146,7 +146,7 @@ public sealed class SubmitShipmentCommandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
@@ -190,7 +190,7 @@ public sealed class SubmitShipmentCommandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
